Bug 1329002 - Remove jobs model
