+++ tests/e2e/test_jobs_loaded.py
@@ -4,10 +4,10 @@ from webtest import TestApp
 from treeherder.config.wsgi import application
 
 
-def test_pending_job_available(jm, pending_jobs_stored):
+def test_pending_job_available(test_repository, pending_jobs_stored):
     webapp = TestApp(application)
     resp = webapp.get(
-        reverse("jobs-list", kwargs={"project": jm.project})
+        reverse("jobs-list", kwargs={"project": test_repository.name})
     )
     jobs = resp.json
 
@@ -16,10 +16,10 @@ def test_pending_job_available(jm, pending_jobs_stored):
     assert jobs['results'][0]['state'] == 'pending'
 
 
-def test_running_job_available(jm, running_jobs_stored):
+def test_running_job_available(test_repository, running_jobs_stored):
     webapp = TestApp(application)
     resp = webapp.get(
-        reverse("jobs-list", kwargs={"project": jm.project})
+        reverse("jobs-list", kwargs={"project": test_repository.name})
     )
     jobs = resp.json
 
@@ -28,10 +28,10 @@ def test_running_job_available(jm, running_jobs_stored):
     assert jobs['results'][0]['state'] == 'running'
 
 
-def test_completed_job_available(jm, completed_jobs_stored):
+def test_completed_job_available(test_repository, completed_jobs_stored):
     webapp = TestApp(application)
     resp = webapp.get(
-        reverse("jobs-list", kwargs={"project": jm.project})
+        reverse("jobs-list", kwargs={"project": test_repository.name})
     )
     jobs = resp.json
 
@@ -39,7 +39,8 @@ def test_completed_job_available(jm, completed_jobs_stored):
     assert jobs['results'][0]['state'] == 'completed'
 
 
-def test_pending_stored_to_running_loaded(jm, pending_jobs_stored, running_jobs_stored):
+def test_pending_stored_to_running_loaded(test_repository, pending_jobs_stored,
+                                          running_jobs_stored):
     """
     tests a job transition from pending to running
     given a loaded pending job, if I store and load the same job with status running,
@@ -47,7 +48,7 @@ def test_pending_stored_to_running_loaded(jm, pending_jobs_stored, running_jobs_
     """
     webapp = TestApp(application)
     resp = webapp.get(
-        reverse("jobs-list", kwargs={"project": jm.project})
+        reverse("jobs-list", kwargs={"project": test_repository.name})
     )
     jobs = resp.json
 
@@ -55,13 +56,14 @@ def test_pending_stored_to_running_loaded(jm, pending_jobs_stored, running_jobs_
     assert jobs['results'][0]['state'] == 'running'
 
 
-def test_finished_job_to_running(jm, completed_jobs_stored, running_jobs_stored):
+def test_finished_job_to_running(test_repository, completed_jobs_stored,
+                                 running_jobs_stored):
     """
     tests that a job finished cannot change state
     """
     webapp = TestApp(application)
     resp = webapp.get(
-        reverse("jobs-list", kwargs={"project": jm.project})
+        reverse("jobs-list", kwargs={"project": test_repository.name})
     )
     jobs = resp.json
 
@@ -69,14 +71,15 @@ def test_finished_job_to_running(jm, completed_jobs_stored, running_jobs_stored)
     assert jobs['results'][0]['state'] == 'completed'
 
 
-def test_running_job_to_pending(jm, running_jobs_stored, pending_jobs_stored):
+def test_running_job_to_pending(test_repository, running_jobs_stored,
+                                pending_jobs_stored):
     """
     tests that a job transition from pending to running
     cannot happen
     """
     webapp = TestApp(application)
     resp = webapp.get(
-        reverse("jobs-list", kwargs={"project": jm.project})
+        reverse("jobs-list", kwargs={"project": test_repository.name})
     )
     jobs = resp.json
