+++ tests/e2e/test_perf_ingestion.py
@@ -7,7 +7,7 @@ from treeherder.perf.models import (PerformanceDatum,
                                     PerformanceSignature)
 
 
-def test_post_perf_artifact(jobs_ds, test_repository, failure_classifications,
+def test_post_perf_artifact(test_repository, failure_classifications,
                             result_set_stored, mock_post_json):
     PerformanceFramework.objects.get_or_create(name='cheezburger', enabled=True)
 
@@ -99,7 +99,7 @@ def test_post_perf_artifact_revision_hash(test_repository, failure_classificatio
     assert PerformanceDatum.objects.all().count() == 3
 
 
-def test_post_perf_artifact_multiple(jobs_ds, test_repository,
+def test_post_perf_artifact_multiple(test_repository,
                                      failure_classifications,
                                      result_set_stored, mock_post_json):
     PerformanceFramework.objects.get_or_create(name='cheezburger', enabled=True)
