--- treeherder/model/derived/__init__.py
@@ -1,2 +1,0 @@
-from .base import *
-from .jobs import *
